sms: fix time zone handling
This fixes two small bugs related to parsing timestamp PDUs:

In (Timestamp).PDU(), generating the time zone octet accidentally
used integer arithmetic to calculate the GMT offset. This causes
problems for time zones like +08:15, which resulted in an offset
of 8 hours (and not 8.25 hours).

In (*Timestamp).ReadFrom(), the wrong bit was used to identify
negative time zones: Negative time zones are actually indicated
by bit 3 (the fourth bit, 0x08, 0b0000_1000) set to one, this code
previously looked at bit 2 (0x04, 0b0000_0100). This also surfaced
an issue with the test fixture for TestSmsDeliverReadFromGsm7_2()
where the correct time zone is GMT-03:00 (Moskow time, which
matches the TP-UD payload contained in that message).